requirements.txt
Flask==3.0.3
gunicorn==21.2.0

openai==1.52.0

PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
Werkzeug==3.0.3

reportlab==4.2.2